+++ src/kafka/types/IResponseValue.ts
@@ -4,7 +4,7 @@ import { ITransactionContext } from './ITransactionContext'
 
 export interface IResponseValue {
   id: string
-  jobId: string
+  jobUUID: string
   txContext?: ITransactionContext
   receipt?: IReceipt
   errors?: IError[]