+++ src/stubs/index.d.ts
@@ -6875,8 +6875,8 @@ export namespace tx {
         /** TxResponse id */
         id?: (string|null);
 
-        /** TxResponse jobId */
-        jobId?: (string|null);
+        /** TxResponse jobUUID */
+        jobUUID?: (string|null);
 
         /** TxResponse contextLabels */
         contextLabels?: ({ [k: string]: string }|null);
@@ -6909,8 +6909,8 @@ export namespace tx {
         /** TxResponse id. */
         public id: string;
 
-        /** TxResponse jobId. */
-        public jobId: string;
+        /** TxResponse jobUUID. */
+        public jobUUID: string;
 
         /** TxResponse contextLabels. */
         public contextLabels: { [k: string]: string };